works/api-server/requirements.txt
# ── API Server ────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.2
boto3==1.38.12
psycopg2-binary==2.9.10
pytesseract==0.3.13
Pillow==11.2.1
pdf2image==1.17.0
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.9.1

# ── Testing ───────────────────────────────────
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
moto[s3]==5.1.4
